==> scripts/analyses/1.Merging.Rmd <==
---
title: 'EDGI protocol: Merging data'
author: "Helena Davies"
date: "24/08/2022"
output: html_document
---

```{r Setup, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE,
  comment = '',
  prompt = FALSE,
  cache = FALSE
  )
```

Clear global environment prior to initiation
```{r Clear global environment}
remove(list = ls())
```

Read in file with path to ilovedata channel on Teams
Ensure that your credentials directory is correctly located
```{r Read in paths file}
source(file = "../credentials/paths.R")
```

Add the add_numeric function - used to convert character variables into numeric variables.
Add the sumscores function - used to generate sumscores
Add the package_check function - used to install and load dependencies
```{r Read in functions}
source(file = paste0(ilovedata_scripts, "functions/add_numeric.R"))
source(file = paste0(ilovedata_scripts, "functions/remove_duplicates.R"))
source(file = paste0(ilovedata_scripts, "functions/sumscores.R"))
source(file = paste0(ilovedata_scripts, "functions/package_check.R"))
source(file = paste0(ilovedata_scripts, "functions/imp_check.R"))
source(file = paste0(ilovedata_scripts, "functions/recode_check.R"))
```

Use package_check to install and load dependencies
Load tidyverse last
```{r Install load dependencies}
packages <- c("summarytools", "sjlabelled", "Amelia", "gtsummary", "tidyverse")
package_check(packages)
```

Retrieve recent date
We are using the recent date to save files with paste0() as an extension to not overwrite old versions
```{r Recent date}
date <- Sys.Date()
date
```

# Read in gender 
```{r read in gender sex and sexuality data}
sex_gender_sexuality <- readRDS(file = paste0(filepath_cleaned_data,
                                     "/demographics/sex_gender_sexuality_edgi_clean.rds"))
  
# Check
sex_gender_sexuality %>%
  colnames()

sex_gender_sexuality %>%
  dim()

# Select variables we need 
sex_gender_sexuality <- sex_gender_sexuality %>%
  select(ID,
         sample,
         sex = dem.sex,
         demographics.what_gender_do_you_identify_with = dem.what_gender_do_you_identify_with,
         sexuality = dem.what_is_your_sexual_orientation,
         dem.do_you_identify_as_transgender)

# Change to character
sex_gender_sexuality$demographics.what_gender_do_you_identify_with <- as.character(sex_gender_sexuality$demographics.what_gender_do_you_identify_with)

sex_gender_sexuality$sexuality <- as.character(sex_gender_sexuality$sexuality)

sex_gender_sexuality$sex <- as.character(sex_gender_sexuality$sex)

sex_gender_sexuality <- sex_gender_sexuality %>%
  mutate(gender =
           case_when(
                     demographics.what_gender_do_you_identify_with == "Prefer to self-define (please tell us more):" ~ "Prefer to self-define",
                     
                     demographics.what_gender_do_you_identify_with == "Male" ~ "Man",
                     demographics.what_gender_do_you_identify_with == "Female" ~ "Woman",
                     TRUE ~ demographics.what_gender_do_you_identify_with
                     )
         )

# Freq
sex_gender_sexuality %>%
  freq(gender)
```

# Read in 'Ethnicity'
```{r read in ethnicity}
ethnicity <- readRDS(file = paste0(filepath_cleaned_data,
                                     "/demographics/ethnicity_edgi_clean.rds"))
  
# Check
ethnicity %>%
  colnames()

ethnicity %>%
  dim()

# Select variables we need 
ethnicity <- ethnicity %>%
  select(ID,
         ethnicity = dem.what_is_your_ethnic_origin)

# Change to character
ethnicity$ethnicity <- as.character(ethnicity$ethnicity)

# Freq
ethnicity %>%
  freq(ethnicity)
```

# Education
```{r read in education}
education <- readRDS(file = paste0(filepath_cleaned_data,
                                     "/demographics/highest_education_edgi_clean.rds"))
  
# Check
education %>%
  colnames()

education %>%
  dim()

# Select variables we need 
education <- education %>%
  select(ID,
         dem.highest_education_finegrained)

# Freq
education %>%
  freq(dem.highest_education_finegrained)
```

```{r read in years of education}
years_of_education <- readRDS(file = paste0(filepath_cleaned_data,
                                     "/demographics/years_of_education_edgi_clean.rds"))
  
# Check
years_of_education %>%
  colnames()

years_of_education %>%
  dim()

# Select variables we need 
years_of_education <- years_of_education %>%
  select(ID,
         dem.university_education_attend_school.txt)

# freq
years_of_education %>%
  freq(dem.university_education_attend_school.txt)
```

# Employment
```{r read in employment}
employment <- readRDS(file = paste0(filepath_cleaned_data,
                                     "/demographics/employment_edgi_clean.rds"))
  
# Check
employment %>%
  colnames()

employment %>%
  dim()

# Needs fixing
# Select variables we need 
employment <- employment %>%
  select(ID,
         dem.what_is_your_current_employment_status_numeric)

employment <- employment %>%
  mutate(dem.what_is_your_current_employment_status =
           case_when(dem.what_is_your_current_employment_status_numeric == -999 ~ "Prefer not to answer",
                     dem.what_is_your_current_employment_status_numeric == 1 ~ "In paid employment or self-employed",
                     dem.what_is_your_current_employment_status_numeric == 2 ~ "Retired",
                     dem.what_is_your_current_employment_status_numeric == 3 ~ "Looking after home and/or family",
                     dem.what_is_your_current_employment_status_numeric == 4 ~ "Unable to work because of sickness or disability",
                     dem.what_is_your_current_employment_status_numeric == 5 ~ "Unemployed",
                     dem.what_is_your_current_employment_status_numeric == 6 ~ "Doing unpaid or voluntary work",
                     dem.what_is_your_current_employment_status_numeric == 7 ~ "Full-time or part-time student",
                     
                     dem.what_is_your_current_employment_status_numeric == 9 ~ "Other"
                     
                     ))

employment %>% 
  freq(dem.what_is_your_current_employment_status_numeric)
```

# Relationship status
```{r read in relationship status}
# Needs fixing
relationship_status <- readRDS(file = paste0(filepath_cleaned_data,
                                     "/demographics/marital_status_edgi_clean.rds"))
  
# Check
relationship_status %>%
  colnames()

relationship_status %>%
  dim()

# Select variables we need 
relationship_status <- relationship_status %>%
  select(ID,
         relationship_status = dem.what_is_your_current_maritalrelationship_status)

# Convert to character
relationship_status$relationship_status <- as.character(relationship_status$relationship_status)

# Freq
relationship_status %>%
  freq(relationship_status)
```

# Smoking status 
Note: This script needs further reviewing. HLD has done a quick review.
```{r read in smoking status}
smoking_status <- readRDS(file = paste0(filepath_cleaned_data,
                                     "/demographics/smoking_pack_year_edgi_clean.rds"))
  
# Check
smoking_status %>%
  colnames()

smoking_status %>%
  dim()

# Select variables we need 
smoking_status <- smoking_status %>%
  select(ID,
         dem.pack_year)

smoking_status %>% 
  freq(dem.pack_year)

# Missing information version
smoking_status <- smoking_status %>%
  mutate(dem.pack_year_missing_info =
    case_when(
    dem.pack_year == -666 ~ "Implausible",
    dem.pack_year == -999 ~ "Prefer not to answer",
    dem.pack_year == -888 ~ "Don't know",
    dem.pack_year == -777 ~ "Seen but not answered",
    !is.na(dem.pack_year) ~ "Answered",
    TRUE ~ NA_character_
    )
  )

smoking_status <- smoking_status %>%
  mutate(dem.pack_year =
    case_when(
      dem.pack_year < 0 ~ NA_real_, # For the continuous version, we need to assign all implausible/PNTA/DK values as NA_real_
    !is.na(dem.pack_year) ~ dem.pack_year,
    TRUE ~ NA_real_
    )
  )


# Check
smoking_status %>%
  freq(dem.pack_year_missing_info)
```

# Treatment
Note: For some reason, this is saved in the MHD even though it's part of the ED100K screening questionnaire.
```{r read in treatment data}
treatment <- readRDS(file = paste0(filepath_edgi_raw,
                                     "mhd_edgi.rds"))
  
# Check
treatment %>%
  colnames()

treatment %>%
  dim()

# Select variables we need 
treatment <- treatment %>%
  select(ID = externalDataReference,
         mhd.eating_disorder_received_treatment
         
         )
```

# Age
```{r read in age data}
age <- readRDS(file = paste0(filepath_cleaned_data,
                                     "/demographics/age_edgi_clean.rds"))
  
# Check
age %>%
  colnames()

age %>%
  dim()

# Select variables we need 
age <- age %>%
  select(ID,
         dem.how_old_are_you_now.txt,
         dem.dob_age,
         dem.age_category)

# Check
age %>%
  freq(dem.how_old_are_you_now.txt)

age %>%
  freq(dem.dob_age)

age <- age %>%
  mutate(age =
    case_when(
      dem.dob_age < 0 ~ NA_real_, # For the continuous version, we need to assign all implausible/PNTA/DK values as NA_real_
     dem.how_old_are_you_now.txt < 0 ~ NA_real_,
      !is.na(dem.dob_age) ~ dem.dob_age,
    !is.na(dem.how_old_are_you_now.txt) ~ dem.how_old_are_you_now.txt,
    TRUE ~ NA_real_
    )
  )

# Check
age %>%
  freq(age)

# Missing information version
age <- age %>%
  mutate(age_missing_info =
    case_when(
      !is.na(age) ~ "Answered",
    dem.dob_age == -666 ~ "Implausible",
    dem.dob_age == -999 ~ "Prefer not to answer",
    dem.dob_age == -888 ~ "Don't know",
    dem.dob_age == -777 ~ "Seen but not answered",
    
    dem.how_old_are_you_now.txt == -666 ~ "Implausible",
    dem.how_old_are_you_now.txt == -999 ~ "Prefer not to answer",
    dem.how_old_are_you_now.txt == -888 ~ "Don't know",
    dem.how_old_are_you_now.txt == -777 ~ "Seen but not answered",
    
    TRUE ~ NA_character_
    )
  )

# Check
age %>%
  freq(age_missing_info)
```

# Eating disorder diagnosis: algorithm-derived
```{r read in eating disorder algorithm data}
eating_disorder_algorithm <- readRDS(file = paste0(filepath_cleaned_data,
                                     "algorithms/ed_algorithms_diagnostics_edgi_clean_040523.rds"))
  
# Check
eating_disorder_algorithm %>%
  colnames()

eating_disorder_algorithm %>%
  dim()

# Select variables we need 
eating_disorder_algorithm <- eating_disorder_algorithm %>%
  select(ID,
         ed.DSM5_AN_binary,
         ed.DSM5_AN_restricting_binary,
         ed.DSM5_AN_binge_purge_binary,
         ed.DSM5_BN_binary,
         ed.DSM5_BED_binary)
```

# Self-report (MHD)
```{r read in mhd data}
# Needs fixing
mhd1 <- readRDS(file = paste0(filepath_cleaned_data,
                                     "clinical/mhd_edgi_clean.rds"))
mhd1 <- mhd1 %>%
  select(-mhd.autism_spectrum_disorder)

mhd2 <- readRDS(file = paste0(filepath_edgi_raw,
                              "dem_edgi.rds"))

mhd2 <- mhd2 %>%
  select(ID = externalDataReference,
         mhd.autism_spectrum_disorder = dem.autism_spectrum_disorder_)

# Merge 
mhd <- dplyr::full_join(mhd1,
                        mhd2,
                        by = "ID")
              
# Check
mhd %>%
  colnames()

mhd %>%
  freq(mhd.gad_numeric)

# Recode personality disorder
mhd <- mhd %>%
  mutate(mhd.personality_disorder_type_diagnosed =
           case_when(
           mhd.personality_disorder_numeric == -777 ~ "Seen but not answered",
           mhd.personality_disorder_numeric == -888 ~ "Don't know",
           mhd.personality_disorder_numeric == -999 ~ "Prefer not to answer",
             mhd.personality_disorder_numeric == 0 ~ "No personality disorder",
           mhd.personality_disorder_diagnosed_numeric == 1 ~ "Paranoid personality disorder",
           mhd.personality_disorder_diagnosed_numeric == 2 ~ "Schizoid personality disorder",
           mhd.personality_disorder_diagnosed_numeric == 3 ~ "Schizotypal personality disorder",
           mhd.personality_disorder_diagnosed_numeric == 4~ "Antisocial personality disorder",
           mhd.personality_disorder_diagnosed_numeric == 5~ "Borderline/Emotionally unstable, personality disorder",
           mhd.personality_disorder_diagnosed_numeric == 6~ "Histrionic personality disorder",
           mhd.personality_disorder_diagnosed_numeric == 7~ "Narcissistic personality disorder",
               mhd.personality_disorder_diagnosed_numeric == 8~ "Avoidant/anxious personality disorder",
                mhd.personality_disorder_diagnosed_numeric == 9~ "Dependent personality disorder",
                 mhd.personality_disorder_diagnosed_numeric == 10~ "Obsessive-compulsive personality disorder",
         mhd.personality_disorder_diagnosed_numeric == 11 ~ "Mixed personality disorder"
         
         )
  )

mhd <- mhd %>%
  mutate(mhd.any_personality_disorder =
           case_when(
             mhd.personality_disorder_numeric == -777 ~ "Seen but not answered",
               mhd.personality_disorder_numeric == -888 ~ "Don't know",
           mhd.personality_disorder_numeric == -999 ~ "Prefer not to answer",
           mhd.personality_disorder_numeric == 0 ~ "No personality disorder",
           mhd.personality_disorder_diagnosed_numeric == 1 |
           mhd.personality_disorder_diagnosed_numeric == 2 |
           mhd.personality_disorder_diagnosed_numeric == 3 |
           mhd.personality_disorder_diagnosed_numeric == 4 |
           mhd.personality_disorder_diagnosed_numeric == 5 |
           mhd.personality_disorder_diagnosed_numeric == 6 |
           mhd.personality_disorder_diagnosed_numeric == 7 |
               mhd.personality_disorder_diagnosed_numeric == 8 |
                mhd.personality_disorder_diagnosed_numeric == 9 |
                 mhd.personality_disorder_diagnosed_numeric == 10 |
         mhd.personality_disorder_diagnosed_numeric == 11 ~ "Personality disorder",
         
         TRUE ~ NA_character_
         
         )
  )

# Select columns
mhd <- mhd %>%
  select(#-endDate,
         #  -startDate,
           -sample)

```

# Somatic comorbidity
```{r read in somatic data}
somatic <- readRDS(file = paste0(filepath_cleaned_data,
                                     "/demographics/disability_illness_edgi_clean.rds"))
  
# Check
somatic %>%
  colnames()

somatic %>%
  dim()

# Select columns
somatic <- somatic %>%
  select(-endDate,
           -startDate,
           -sample)

# Create overarching variables
somatic <- somatic %>%
  mutate(any_cancer_numeric =
           case_when(dem.breast_cancer == "Breast cancer" | 
             dem.lung_cancer == "Lung cancer" |
             dem.stomach_cancer == "Stomach cancer" |
             dem.colon_cancer == "Colon cancer" |
             dem.uterus_cancer ==  "Uterus cancer" |
             dem.prostate_cancer ==  "Prostate cancer" ~ 1,
             
             TRUE ~ 0
             
           )) %>%
  mutate(any_neurological_numeric =
           case_when(dem.epilepsy_or_convulsions == "Epilepsy or convulsions" |
                     dem.migraines == "Migraines" |
                      dem.multiple_sclerosis == "Multiple sclerosis" |
                      dem.parkinsons_disease == "Parkinson's disease" |
                      dem.brain_tumour == "Brain tumour" |
                      dem.severe_memory_loss == "Severe memory loss" ~ 1,
                     
                     TRUE ~ 0
                     )) %>%
  mutate(any_allergy_numeric =
           case_when(
             dem.hay_fever == "Hay fever" |
               dem.drug_allergy == "Drug allergy" |
             dem.food_allergy == "Food allergy" |
             dem.other_allergy == "Other allergy" ~ 1,
             
             TRUE ~ 0
            
           )) %>%
   mutate(bone_joint_problem_numeric =
           case_when(
             dem.ankylosing_spondylitis == "Ankylosing spondylitis" |
              dem.osteoporosis == "Osteoporosis" |
              dem.osteoarthritis == "Osteoarthritis" |
             dem.hypermobility == "Hyper-mobility" |
             dem.rheumatoid_arthritis == "Rheumatoid arthritis" |
             dem.other_arthritis == "Other arthritis"~ 1,
           
               TRUE ~ 0
             )) %>%
  mutate(lung_breathing_problems_numeric =
           case_when(dem.asthma == "Asthma" |
                     dem.emphysema_or_chronic_bronchitis == "Emphysema or chronic bronchitis" ~ 1,
                     TRUE ~ 0
                     )) %>%
  mutate(heart_circulation_problem_numeric =
           case_when(dem.heart_attack_or_angina =="Heart attack or angina" |
                     dem.high_blood_cholesterol == "High blood cholesterol" |
                     dem.high_blood_pressure == "High blood pressure" |
                     dem.atrial_fibrillation == "Atrial fibrillation" |
                     dem.stroke == "Stroke" |
                     dem.pots == "Paroxysmal tachycardia syndrome (PoTS)" ~ 1,
             
                     TRUE ~ 0
                     )) %>%
  mutate(digestive_problem_numeric =
           case_when(dem.crohns_disease == "Crohn's disease" |
                     dem.ulcerative_colitis == "Ulcerative colitis" |
                     dem.coeliac_disease == "Coeliac disease" ~ 1,
                     
                      TRUE ~ 0
                     )) %>%
  mutate(diabetes_numeric =
           case_when(
             dem.diabetes_type_1_early_onset == "Diabetes type 1 (early onset)" |
             dem.diabetes_type_1_late_onset ==  "Diabetes type 1 (late onset)" |
             dem.diabetes_type_2_late_onset  == "Diabetes type 2 (late onset)" |
               dem.pain_due_to_diabetes == "Pain due to diabetes (diabetic neuropathy)" ~ 1,
             
             TRUE ~ 0
                     )) %>%
  mutate(polycystic_ovary_numeric =
           case_when(dem.pcos == "Polycystic ovary syndrome (PCOS)" ~ 1,
                     
                     TRUE ~ 0
                     )) %>%
  mutate(skin_problems_numeric =
           case_when(dem.psoriasis == "Psoriasis" |
                     dem.eczema == "Eczema" |
                     dem.vitiligo == "Vitiligo" |
                     dem.pain_due_to_virus == "Pain due to virus" ~ 1, 
                     
                     TRUE ~ 0
             
           )) %>%
  mutate(thyroid_disease_numeric =
           case_when(
             dem.thyroid_disease == "Thyroid disease" ~ 1,
             
             TRUE ~ 0
           ))
```

# Age at diagnosis
```{r read in age at diagnosis data}
age_at_diagnosis <- readRDS(file = paste0(filepath_cleaned_data,
                                     "/demographics/age_at_diagnosis_edgi_clean.rds"))
  
# Check
age_at_diagnosis %>%
  colnames()

age_at_diagnosis %>%
  dim()
```

# OSFED / other diagnoses
```{r read in OSFED diagnoses}
OSFED_other_eating_disorders <- readRDS(file = paste0(filepath_cleaned_data,
                                     "/algorithms/osfed_dsm5_algorithms_edgi_clean_040523.rds"))
  
# Check
OSFED_other_eating_disorders %>%
  colnames()

OSFED_other_eating_disorders %>%
  dim()
```

# Merge all data
```{r}
dat <- list(sex_gender_sexuality,
                ethnicity,
                education,
                years_of_education,
                employment,
                relationship_status,
                smoking_status,
                treatment,
                age,
                eating_disorder_algorithm,
                mhd,
                somatic,
                OSFED_other_eating_disorders,
                age_at_diagnosis
                ) %>%
  reduce(full_join,
         by = "ID")
```

```{r filter AAN}
# Filter out BN and BED cases from AAN
dat <- dat %>%
  mutate(ed.DSM5_AN_atypical_binary_combined_filtered =
  case_when(ed.DSM5_AN_atypical_binary_combined_filtered == "DSM-5 atypical AN combined and filtered" &
              
             ed.DSM5_BN_binary == "No DSM-5 BN" &
             ed.DSM5_BED_binary == "No DSM-5 BED" ~ "DSM-5 atypical AN combined and filtered" )
  )

dat %>%
  freq(ed.DSM5_AN_atypical_binary_combined_filtered)
```

# Look at the data
```{r view table}
skimr::skim(dat)

dat %>%
  colnames()

dat %>%
  dim()
```

# Save the data
```{r save the merged data}
dat %>% 
  saveRDS(
    file = paste0(filepath_cleaned_data, "merged_dat_040523.rds")
    )

```
