==> scripts/analyses/2.Variable_creation.Rmd <==
---
title: "Demographics"
author: "Christopher Huebel, Helena Davies, Dina Monssen"
date: "2022-08-24"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Clear Global Environment

```{r clear global environment}
rm(list = ls(all.names = TRUE)) 
```

Read in file with path to ilovedata channel on Teams
Ensure that your credentials directory is correctly located
```{r Read in paths file}
source(file = "../credentials/paths.R")
```

Add the add_numeric function - used to convert character variables into numeric variables.
Add the sumscores function - used to generate sumscores
Add the package_check function - used to install and load dependencies
```{r Read in functions}
source(file = paste0(ilovedata_scripts, "functions/add_numeric.R"))
source(file = paste0(ilovedata_scripts, "functions/remove_duplicates.R"))
source(file = paste0(ilovedata_scripts, "functions/sumscores.R"))
source(file = paste0(ilovedata_scripts, "functions/package_check.R"))
source(file = paste0(ilovedata_scripts, "functions/imp_check.R"))
source(file = paste0(ilovedata_scripts, "functions/recode_check.R"))
```

Load packages
```{r Load packages}
library(tidyverse)
library(summarytools)
library(gtsummary)
```

Set current date
```{r current date}
date <- Sys.Date()
```

Load data
```{r Load data}
dat <- read_rds(
  file = paste0(filepath_cleaned_data, "merged_dat_040523.rds")
)
```

Overview variables
```{r Overview variables}
names(dat)
```

# Check response options
```{r check response options}
dat %>%
  freq(ed.DSM5_AN_binary)

dat %>%
  freq(mhd.an_numeric)

dat %>%
  freq(ed.DSM5_BN_binary)

dat %>%
  freq(mhd.bn_numeric)

dat %>%
  freq(ed.DSM5_BED_binary)

dat %>%
  freq(mhd.bed_numeric
       )
```


```{r}
dat <- dat %>%
  mutate(across(contains("numeric"),
                ~case_when(
                  . == -999 ~ "Prefer not to answer",
                  . == -888 ~ "Don't know",
                  . == -777 ~ "Seen but not answered",
                  TRUE ~ as.character(.)
                  )
                )
         )

# Check
dat %>%
  freq(mhd.bed_numeric)
```

# First, filter out ineligible participants
Participants must have an ED, age and sex data.
```{r filter out ineligible participants}
# Filter out age and sex missing
dat <- dat %>% 
  filter(!is.na(age) &
           !is.na(sex))
```

# Self-report and algorithm-derived ED diagnosis
```{r deriving self-report and algorithm-derived ED diagnosis}
dat <- dat %>%
  mutate(
    anorexia_nervosa =
      case_when(
        ed.DSM5_AN_binary == "DSM-5 AN" |
          mhd.an_numeric == 1 ~ "Anorexia nervosa"),
    
    bulimia_nervosa =
      case_when(
        ed.DSM5_BN_binary == "DSM-5 BN" |
          mhd.bn_numeric == 1 ~ "Bulimia nervosa"),
    
    binge_eating_disorder =
           case_when(
             ed.DSM5_BED_binary == "DSM-5 BED" |
               mhd.bed_numeric == 1 ~ "Binge-eating disorder"),
             
      purging_disorder =
        case_when(ed.DSM5_purging_disorder_binary_numeric == 1 |
                    mhd.purging_disorder_numeric == 1 ~ "Purging disorder"),
    
    atypical_anorexia_nervosa =
      case_when(mhd.atypical_an_numeric == 1 |
                  ed.DSM5_AN_atypical_binary_combined_filtered == "DSM-5 atypical AN combined and filtered" ~ "Atypical anorexia nervosa"),
                
      atypical_binge_eating_disorder =
        case_when(mhd.atypical_bed_numeric == 1 |
                   ed.DSM5_BED_atypical_binary_numeric == 1 ~ "Atypical binge-eating disorder"),
                  
      atypical_bulimia_nervosa =
        case_when(ed.DSM5_BN_atypical_binary_numeric == 1 |
                   mhd.atypical_bn_numeric == 1 ~ "Atypical bulimia nervosa"),
      
      night_eating_syndrome = 
        case_when(ed.DSM5_NES_binary_numeric == 1 |
                    mhd.nighteating_syndrome_numeric == 1 ~ "Night eating syndrome"),
      
      pica =
        case_when(mhd.pica_numeric == 1 ~ "Pica"),
      
      avoidant_restrictive_food_intake_disorder =
        case_when(mhd.avoidantrestrictive_food_intake_disorder_numeric == 1~ "ARFID"),
      
       rumination_disorder =
        case_when(mhd.rumination_disorder_numeric == 1 ~ "Rumination disorder"),
      
      )
      
# Check n
dat %>%
  freq(anorexia_nervosa)

dat %>%
  freq(bulimia_nervosa)

dat %>%
  freq(binge_eating_disorder)
```

# Filter out participants with no eating disorder
```{r filter out participants with no eating disorder}
dat <- dat %>%
  filter(!is.na(anorexia_nervosa) |
  !is.na(bulimia_nervosa) |
  !is.na(binge_eating_disorder) |
  !is.na(purging_disorder) |
  !is.na(atypical_anorexia_nervosa) |
  !is.na(atypical_binge_eating_disorder) |
  !is.na(atypical_bulimia_nervosa) |
  !is.na(night_eating_syndrome) |
  !is.na(pica) |
  !is.na(avoidant_restrictive_food_intake_disorder) |
  !is.na(rumination_disorder) |
  
    # Previously we used: "Have you ever suspected that you may have an eating disorder, whether or not you were ever diagnosed?
  # mhd.suspected_eating_disorder_diagnosed_numeric == 1
 
   # Now, we are using: Have you ever received treatment for an eating disorder?
  mhd.eating_disorder_received_treatment == 1 |
    
  # Have you ever been diagnosed with 'one or more of the following eating disorders by a professional, even if you don't have it currently?By professional we mean: any doctor, nurse or person with a specialist training (such as a psychologist, psychiatrist etc.) Please include disorders even if you did not need treatment for them or if you did not agree with the diagnosis: Other eating disorder
    mhd.other_eating_disorder_numeric == 1 |
    
     # Have you ever been diagnosed with 'one or more of the following eating disorders by a professional, even if you don't have it currently?By professional we mean: any doctor, nurse or person with a specialist training (such as a psychologist, psychiatrist etc.) Please include disorders even if you did not need treatment for them or if you did not agree with the diagnosis: Other specified feeding or eating disorder (OSFED or EDNOS)
   mhd.feeding_eating_disorder_numeric == 1
  )
```

```{r numeric deriving self-report and algorithm-derived ED diagnosis}
dat <- dat %>%
  mutate(
    anorexia_nervosa_numeric =
      case_when(ed.DSM5_AN_binary == "DSM-5 AN" |
          mhd.an_numeric == 1 ~ 1,
          
          TRUE ~ 0),
    
    anorexia_nervosa_subtype_numeric =
      case_when(
        ed.DSM5_AN_binge_purge_binary == "DSM-5 AN binge eating/purging"  ~ "Binge-eating/purging",
        
        ed.DSM5_AN_restricting_binary == "DSM-5 AN restricting" ~ "Restricting",
        
        ed.DSM5_AN_binary == "DSM-5 AN" |
          mhd.an_numeric == 1 ~ "Unknown subtype",
        
        TRUE ~ NA_character_),

    bulimia_nervosa_numeric =
      case_when(
        ed.DSM5_BN_binary == "DSM-5 BN" |
          mhd.bn_numeric == 1 ~ 1,
        
         TRUE ~ 0),
    
    binge_eating_disorder_numeric =
           case_when(
             ed.DSM5_BED_binary == "DSM-5 BED" |
               mhd.bed_numeric == 1 ~ 1,
             
             TRUE ~ 0),
             
      purging_disorder_numeric =
        case_when(ed.DSM5_purging_disorder_binary_numeric == 1 |
                    mhd.purging_disorder_numeric == 1 ~ 1,
                  
                  TRUE ~ 0),
    
    atypical_anorexia_nervosa_numeric =
      case_when(mhd.atypical_an_numeric == 1 |
                  ed.DSM5_AN_atypical_binary_combined_filtered == "DSM-5 atypical AN combined and filtered" ~ 1,
              
                TRUE ~ 0),
                
      atypical_binge_eating_disorder_numeric =
        case_when(mhd.atypical_bed_numeric == 1 |
                   ed.DSM5_BED_atypical_binary_numeric == 1 ~ 1,
                  
                  TRUE ~ 0),
                  
      atypical_bulimia_nervosa_numeric =
        case_when(ed.DSM5_BN_atypical_binary_numeric == 1 |
                   mhd.atypical_bn_numeric == 1 ~ 1,
                  
                  TRUE ~ 0),
      
      night_eating_syndrome_numeric = 
        case_when(ed.DSM5_NES_binary_numeric == 1 |
                    mhd.nighteating_syndrome_numeric == 1 ~ 1,
                  
                  TRUE ~ 0),
      
      pica_numeric =
        case_when(mhd.pica_numeric == 1 ~ 1,
                  TRUE ~ 0),
      
      avoidant_restrictive_food_intake_disorder_numeric =
        case_when(mhd.avoidantrestrictive_food_intake_disorder_numeric == 1 ~ 1,
                  TRUE ~ 0),
      
       rumination_disorder_numeric =
        case_when(mhd.rumination_disorder_numeric == 1 ~ 1,
                  
                  TRUE ~ 0),
      
      )
    
```

```{r Mixed presentation primary}
dat <- dat %>%
  mutate(
    mixed_eating_disorders_primary =
      case_when(
        anorexia_nervosa == "Anorexia nervosa" &
        bulimia_nervosa == "Bulimia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" ~ "AN & BN & BED",
        
        bulimia_nervosa == "Bulimia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" ~ "BN & BED",
        
        anorexia_nervosa == "Anorexia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" ~ "AN & BED",
      
        anorexia_nervosa == "Anorexia nervosa" &
        bulimia_nervosa == "Bulimia nervosa" ~ "AN & BN",
       
        anorexia_nervosa == "Anorexia nervosa" ~ "AN",
        
        bulimia_nervosa == "Bulimia nervosa" ~ "BN",
        
        binge_eating_disorder == "Binge-eating disorder" ~ "BED"
      )
  )

# Check n
dat %>%
  freq(mixed_eating_disorders_primary
       )
```

```{r Mixed presentation primary plus purging and AAN}
dat <- dat %>%
  mutate(
    mixed_eating_disorders_all =
      case_when(
        anorexia_nervosa == "Anorexia nervosa" &
        bulimia_nervosa == "Bulimia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" &
        purging_disorder == "Purging disorder" &
        atypical_anorexia_nervosa == "Atypical anorexia nervosa"
          ~ "All",
        
        anorexia_nervosa == "Anorexia nervosa" &
        bulimia_nervosa == "Bulimia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" &
        purging_disorder == "Purging disorder" 
          ~ "AN & BN & BED & PUR",
        
        anorexia_nervosa == "Anorexia nervosa" &
        bulimia_nervosa == "Bulimia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" &
        atypical_anorexia_nervosa == "Atypical anorexia nervosa"
          ~ "AN & BN & BED & AAN",
        
        anorexia_nervosa == "Anorexia nervosa" &
        bulimia_nervosa == "Bulimia nervosa" &
        purging_disorder == "Purging disorder" &
        atypical_anorexia_nervosa == "Atypical anorexia nervosa"
          ~ "AN & BN & PUR & AAN",
        
        anorexia_nervosa == "Anorexia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" &
        purging_disorder == "Purging disorder" &
        atypical_anorexia_nervosa == "Atypical anorexia nervosa"
          ~ "AN & BED & PUR & AAN",
        
        bulimia_nervosa == "Bulimia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" &
        purging_disorder == "Purging disorder" &
        atypical_anorexia_nervosa == "Atypical anorexia nervosa"
          ~ "BN & BED & PUR & AAN",
        
       binge_eating_disorder == "Binge-eating disorder" &
        purging_disorder == "Purging disorder" &
        atypical_anorexia_nervosa == "Atypical anorexia nervosa"
          ~ "BED & PUR & AAN",
        
        anorexia_nervosa == "Anorexia nervosa" &
        purging_disorder == "Purging disorder" &
        atypical_anorexia_nervosa == "Atypical anorexia nervosa"
          ~ "AN & PUR & AAN",
        
       anorexia_nervosa == "Anorexia nervosa" &
        bulimia_nervosa == "Bulimia nervosa" &
        atypical_anorexia_nervosa == "Atypical anorexia nervosa"
          ~ "AN & BN & AAN",
        
        anorexia_nervosa == "Anorexia nervosa" &
        bulimia_nervosa == "Bulimia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" 
          ~ "AN & BN & BED",
        
        bulimia_nervosa == "Bulimia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" &
        purging_disorder == "Purging disorder" 
          ~ "BN & BED & PUR",
       
        anorexia_nervosa == "Anorexia nervosa" &
         purging_disorder == "Purging disorder" 
          ~ "AN & PUR",
       
       anorexia_nervosa == "Anorexia nervosa" &
         atypical_anorexia_nervosa == "Atypical anorexia nervosa"
          ~ "AN & AAN",
       
        bulimia_nervosa == "Bulimia nervosa" &
        purging_disorder == "Purging disorder" 
          ~ "BN & PUR",
       
        bulimia_nervosa == "Bulimia nervosa" &
         atypical_anorexia_nervosa == "Atypical anorexia nervosa"
          ~ "BN & AAN",
       
        binge_eating_disorder == "Binge-eating disorder" &
        purging_disorder == "Purging disorder" 
          ~ "BED & PUR",
       
        binge_eating_disorder == "Binge-eating disorder" &
        atypical_anorexia_nervosa == "Atypical anorexia nervosa"
          ~ "BED & AAN",
       
      #   purging_disorder == "Purging disorder" &
      #  atypical_anorexia_nervosa == "Atypical anorexia nervosa"
      #    ~ "PUR & AAN",
       
        bulimia_nervosa == "Bulimia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" ~ "BN & BED",
        
        anorexia_nervosa == "Anorexia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" ~ "AN & BED",
      
        anorexia_nervosa == "Anorexia nervosa" &
        bulimia_nervosa == "Bulimia nervosa" ~ "AN & BN",
       
        anorexia_nervosa == "Anorexia nervosa" ~ "AN",
        
        bulimia_nervosa == "Bulimia nervosa" ~ "BN",
        
        binge_eating_disorder == "Binge-eating disorder" ~ "BED",
       
        purging_disorder == "Purging disorder" ~ "PUR",
       
        atypical_anorexia_nervosa == "Atypical anorexia nervosa" ~ "AAN"
      )
  )

# Check n
dat %>%
  freq(mixed_eating_disorders_all
       )

# Should be 0
dat %>% 
  filter(is.na(mixed_eating_disorders_all) &
         is.na(anorexia_nervosa_numeric) &
          is.na(bulimia_nervosa_numeric) &
          is.na(binge_eating_disorder_numeric) &
          is.na(purging_disorder_numeric) &
          is.na(atypical_anorexia_nervosa_numeric)) %>%
  nrow()
```
      
# Checking missing data
```{r check missing}
# All missing = 1,406 (was 3,222)
dat %>%
  filter(is.na(mixed_eating_disorders_primary) &
           is.na(purging_disorder) &
           is.na(atypical_anorexia_nervosa) &
         is.na(atypical_binge_eating_disorder) &
         is.na(atypical_bulimia_nervosa) &
         is.na(night_eating_syndrome) &
         is.na(pica) &
         is.na(avoidant_restrictive_food_intake_disorder) &
         is.na(rumination_disorder)
         ) %>%
           select(
            # Other self-report
           mhd.other_eating_disordertext.txt,                              
           mhd.suspected_eating_disorder_diagnosed_numeric,                 
           mhd.eating_disorder_received_treatment_numeric,                     
            mhd.feeding_eating_disorder_numeric,                                
           mhd.other_eating_disorder_numeric,                              
           mhd.none_eds_numeric,                                             
           mhd.dont_know_eds_numeric
           )

# All missing = 1,406 (was 3,222)
dat %>%
  filter(is.na(mixed_eating_disorders_primary) &
           is.na(purging_disorder) &
           is.na(atypical_anorexia_nervosa) &
         is.na(atypical_binge_eating_disorder) &
         is.na(atypical_bulimia_nervosa) &
         is.na(night_eating_syndrome) &
         is.na(pica) &
         is.na(avoidant_restrictive_food_intake_disorder) &
         is.na(rumination_disorder)
         ) %>%
           select(
            # Other self-report
           mhd.other_eating_disordertext.txt,                              
           mhd.suspected_eating_disorder_diagnosed_numeric,                 
           mhd.eating_disorder_received_treatment_numeric,                     
            mhd.feeding_eating_disorder_numeric,                                
           mhd.other_eating_disorder_numeric,                              
           mhd.none_eds_numeric,                                             
           mhd.dont_know_eds_numeric
           )

# Missing for received treatment = 0 (was 951)
dat %>%
  filter(is.na(mixed_eating_disorders_primary) &
           is.na(purging_disorder) &
           is.na(atypical_anorexia_nervosa) &
         is.na(atypical_binge_eating_disorder) &
         is.na(atypical_bulimia_nervosa) &
         is.na(night_eating_syndrome) &
         is.na(pica) &
         is.na(avoidant_restrictive_food_intake_disorder) &
         is.na(rumination_disorder) &
          is.na(mhd.eating_disorder_received_treatment_numeric)
         ) %>%
  nrow()


# Missing for received treatment and suspected eating disorder = 0 (was 951)
dat %>%
  filter(is.na(mixed_eating_disorders_primary) &
           is.na(purging_disorder) &
           is.na(atypical_anorexia_nervosa) &
         is.na(atypical_binge_eating_disorder) &
         is.na(atypical_bulimia_nervosa) &
         is.na(night_eating_syndrome) &
         is.na(pica) &
         is.na(avoidant_restrictive_food_intake_disorder) &
         is.na(rumination_disorder) &
          is.na(mhd.eating_disorder_received_treatment_numeric) &
           is.na(mhd.suspected_eating_disorder_diagnosed_numeric)
         ) %>%
  nrow()

# Missing for received treatment and suspected eating disorder & text response = 0 (was 951)
dat %>%
  filter(is.na(mixed_eating_disorders_primary) &
           is.na(purging_disorder) &
           is.na(atypical_anorexia_nervosa) &
         is.na(atypical_binge_eating_disorder) &
         is.na(atypical_bulimia_nervosa) &
         is.na(night_eating_syndrome) &
         is.na(pica) &
         is.na(avoidant_restrictive_food_intake_disorder) &
         is.na(rumination_disorder) &
          is.na(mhd.eating_disorder_received_treatment_numeric) &
           is.na(mhd.suspected_eating_disorder_diagnosed_numeric) &
           is.na(mhd.other_eating_disordertext.txt)
         ) %>%
  nrow()

# Missing for received treatment and suspected eating disorder & text response  & mhd.feeding_eating_disorder_numeric = 0 (was 750)
dat %>%
  filter(is.na(mixed_eating_disorders_primary) &
           is.na(purging_disorder) &
           is.na(atypical_anorexia_nervosa) &
         is.na(atypical_binge_eating_disorder) &
         is.na(atypical_bulimia_nervosa) &
         is.na(night_eating_syndrome) &
         is.na(pica) &
         is.na(avoidant_restrictive_food_intake_disorder) &
         is.na(rumination_disorder) &
          is.na(mhd.eating_disorder_received_treatment_numeric) &
           is.na(mhd.suspected_eating_disorder_diagnosed_numeric) &
           is.na(mhd.other_eating_disordertext.txt) &
           is.na(mhd.feeding_eating_disorder_numeric)
         ) %>%
  nrow()

# Missing for received treatment and suspected eating disorder & text response  & mhd.feeding_eating_disorder_numeric & AGE = 0 (was 265)
dat %>%
  filter(is.na(mixed_eating_disorders_primary) &
           is.na(purging_disorder) &
           is.na(atypical_anorexia_nervosa) &
         is.na(atypical_binge_eating_disorder) &
         is.na(atypical_bulimia_nervosa) &
         is.na(night_eating_syndrome) &
         is.na(pica) &
         is.na(avoidant_restrictive_food_intake_disorder) &
         is.na(rumination_disorder) &
          is.na(mhd.eating_disorder_received_treatment_numeric) &
           is.na(mhd.suspected_eating_disorder_diagnosed_numeric) &
           is.na(mhd.other_eating_disordertext.txt) &
           is.na(mhd.feeding_eating_disorder_numeric) &
           is.na(dem.age_category) &
           is.na(age)
         ) %>%
  nrow()
```

# Dataset creation: Demographics
```{r demographics}
exclude_columns_from_pivoting <-
  c(
    "sex",
    "age",
    "age_missing_info",
    "gender",
    "dem.do_you_identify_as_transgender",
    "sexuality",
    "ethnicity",
    "dem.highest_education_finegrained",
    "dem.what_is_your_current_employment_status",
    "relationship_status",
    "dem.pack_year",
    "dem.pack_year_missing_info",
    "mhd.eating_disorder_received_treatment"
  )

# Exclusive
dat_demographics <- dat %>%
  select(
   # mixed_eating_disorders_primary,
    mixed_eating_disorders_all,
    sex,
    age,
    age_missing_info,
    gender,
    dem.do_you_identify_as_transgender,
    sexuality,
    ethnicity,
    dem.highest_education_finegrained,
    dem.what_is_your_current_employment_status,
    relationship_status,
    dem.pack_year,
    dem.pack_year_missing_info,
    mhd.eating_disorder_received_treatment
  )

# Long
dat_demographics_long <- dat %>%
  select(
    all_of(exclude_columns_from_pivoting),
  # Disorder
    anorexia_nervosa,
    bulimia_nervosa,
    binge_eating_disorder,
    purging_disorder,
    atypical_anorexia_nervosa
    ) %>%
  pivot_longer(
    -all_of(exclude_columns_from_pivoting),
#    names_to = "Old_variable",
    values_to = "Disorder"
    ) %>%
  select(
    -name
  )

dat_demographics_long %>%
  head()
```


# Dataset creation: Age distribution of EDGI UK sample by age range and sex
```{r select age distribution data}
# Exclusive
age_sex_distribution <- dat %>%
  select(
   # mixed_eating_disorders_primary,
    mixed_eating_disorders_all,
         age, 
         sex)

# Long format
exclude_columns_from_pivoting <-
  c(
    "age",
    "sex"
    )

age_sex_distribution_long <- dat %>%
  select(
    all_of(exclude_columns_from_pivoting),
         
        # Disorder
        anorexia_nervosa,
        bulimia_nervosa,
        binge_eating_disorder,
    purging_disorder,
    atypical_anorexia_nervosa
    
        ) %>%
  pivot_longer(
    -exclude_columns_from_pivoting,
#    names_to = "Old_variable",
    values_to = "Disorder"
    ) %>%
  select(
    -name
  )

age_sex_distribution_long
```

# Dataset creation: Self-report and algorithm derived eating disorder diagnosis distribution in EDGI UK sample, split by sex
```{r select age distribution data}
# Exclusive
selfreport_algo_sex <- dat %>%
  select(
         sex,
         # Algorithm-derived
         ed.DSM5_AN_binary,
         ed.DSM5_AN_restricting_binary,
         ed.DSM5_AN_binge_purge_binary,
         ed.DSM5_BED_binary, 
         ed.DSM5_BN_binary,
        ed.DSM5_purging_disorder_binary,
        ed.DSM5_AN_atypical_binary_combined_filtered,
        ed.DSM5_BED_atypical_binary,
        ed.DSM5_BN_atypical_binary,
        ed.DSM5_NES_binary,
        
         # Self-report
         mhd.an_numeric,                                                    
         mhd.atypical_an_numeric,                                  
         mhd.bn_numeric,                                                
         mhd.bed_numeric,                                                
         mhd.atypical_bn_numeric,                                         
         mhd.atypical_bed_numeric,                                         
         mhd.purging_disorder_numeric,                                
         mhd.nighteating_syndrome_numeric,                                 
         mhd.pica_numeric,                                      
         mhd.avoidantrestrictive_food_intake_disorder_numeric,              
         mhd.rumination_disorder_numeric,                          
         mhd.feeding_eating_disorder_numeric,                              
         mhd.other_eating_disorder_numeric)
         
# Long format
exclude_columns_from_pivoting <-
  c(
    "sex",
     "ed.DSM5_AN_binary",
     "ed.DSM5_AN_restricting_binary",
     "ed.DSM5_AN_restricting_binary",
     "ed.DSM5_BED_binary", 
     "ed.DSM5_BN_binary",
      "ed.DSM5_purging_disorder_binary",
      "ed.DSM5_AN_atypical_binary_combined_filtered",
      "ed.DSM5_BED_atypical_binary",
      "ed.DSM5_BN_atypical_binary",
      "ed.DSM5_NES_binary",
      
      # Self-report
     "mhd.an_numeric",                                                    
     "mhd.atypical_an_numeric",                                  
     "mhd.bn_numeric",                                                
     "mhd.bed_numeric",                                                
     "mhd.atypical_bn_numeric",                                         
     "mhd.atypical_bed_numeric",                                         
     "mhd.purging_disorder_numeric",                                
     "mhd.nighteating_syndrome_numeric",                                 
     "mhd.pica_numeric",                                      
     "mhd.avoidantrestrictive_food_intake_disorder_numeric",              
     "mhd.rumination_disorder_numeric",                          
     "mhd.feeding_eating_disorder_numeric",                              
     "mhd.other_eating_disorder_numeric"
  )

selfreport_algo_sex_long <- dat %>%
  select(
    all_of(exclude_columns_from_pivoting),
    # Disorder
        anorexia_nervosa,
        bulimia_nervosa,
        binge_eating_disorder,
    purging_disorder,
    atypical_anorexia_nervosa
        ) %>%
  pivot_longer(
    -exclude_columns_from_pivoting,
#    names_to = "Old_variable",
    values_to = "Disorder"
    ) %>%
  select(
    -name
  )

selfreport_algo_sex_long
```

```{r Psychiatric comorbidities}
exclude_columns_from_pivoting <-
  c(
    "mhd.depression_and_anxiety",
    "mhd.depressive_disorders",
    "mhd.anxiety_disorders",   
    "mhd.ptsd_numeric",
    
    "mhd.obsessive_compulsive_disorders_numeric",
    
    "mhd.any_personality_disorder",
    
    "mhd.bipolar_and_schizophrenia",
    "mhd.psychotic_disorders",
    "mhd.bipolar_disorder_numeric",
    
    "mhd.autism_spectrum_disorder",
    "mhd.addadhd_numeric"
  )


# mixed 
dat_psych_dx <- dat %>%
  select(
    #mixed_eating_disorders_primary,
    mixed_eating_disorders_all,
    
    all_of(exclude_columns_from_pivoting)
  )

dat_psych_dx_longer <- dat %>%
  select(
    
    all_of(exclude_columns_from_pivoting),
    
    # Diagnosis
    anorexia_nervosa,
    bulimia_nervosa,
    binge_eating_disorder,
     purging_disorder,
    atypical_anorexia_nervosa
    
  ) %>%
  pivot_longer(
    -exclude_columns_from_pivoting,
#    names_to = "Old_variable",
    values_to = "Disorder"
    ) %>%
  select(
    -name
  )


dat_psych_dx_longer %>%
  head()
```


```{r Somatic dx exlcude from pivoting}
exclude_columns_from_pivoting <-
  c(
    "dem.epilepsy_or_convulsions"                                        
 ,"dem.migraines"                                                      
 ,"dem.multiple_sclerosis"                                             
 ,"dem.parkinsons_disease"                                             
 ,"dem.severe_memory_loss"                                             
                       
 ,"dem.hay_fever"                                                      
 ,"dem.drug_allergy"                                                   
 ,"dem.food_allergy"                                                   
 ,"dem.other_allergy"                                                  
 ,"dem.osteoporosis"                                                   
 ,"dem.osteoarthritis"                                                 
 ,"dem.rheumatoid_arthritis"                                           
 ,"dem.other_arthritis"                                                
                                           
 ,"dem.asthma"                                                         
 ,"dem.emphysema_or_chronic_bronchitis"                                
 ,"dem.heart_attack_or_angina"                                         
 ,"dem.high_blood_cholesterol"                                         
 ,"dem.high_blood_pressure"                                            
 ,"dem.atrial_fibrillation"                                            
 ,"dem.stroke"                                                         
                                         
 ,"dem.crohns_disease"                                                 
 ,"dem.ulcerative_colitis"                                             
 ,"dem.coeliac_disease" 
  ,"dem.diabetes_type_1_early_onset"
,"dem.diabetes_type_1_late_onset"
,"dem.diabetes_type_2_late_onset"
,"dem.pain_due_to_diabetes"
 ,"dem.pain_due_to_virus"                                              
                               
 ,"dem.breast_cancer"                                                  
 ,"dem.lung_cancer"                                                    
 ,"dem.stomach_cancer"                                                 
 ,"dem.colon_cancer"                                                   
 ,"dem.uterus_cancer"                                                  
 ,"dem.prostate_cancer"

 ,"dem.psoriasis"
 ,"dem.vitiligo"
 ,"dem.eczema"
 ,"dem.thyroid_disease"


  )
```

```{r Somatic comorbidities}
# mixed 
dat_somatic_dx <- dat %>%
  select(
   # mixed_eating_disorders_primary,
    mixed_eating_disorders_all,
    all_of(exclude_columns_from_pivoting)
  )

# Longer 
dat_somatic_dx_longer <- dat %>%
  select(
    
    all_of(exclude_columns_from_pivoting),
    
    # Diagnosis
    anorexia_nervosa,
    bulimia_nervosa,
    binge_eating_disorder,
    purging_disorder,
    atypical_anorexia_nervosa
  ) %>%
  pivot_longer(
    -exclude_columns_from_pivoting,
#    names_to = "Old_variable",
    values_to = "Disorder"
    ) %>%
  select(
    -name
  )


dat_somatic_dx_longer %>%
  head()
```

```{r broad somatic dx exlcude from pivoting}
exclude_broad_columns_from_pivoting <-
  c(
    "any_cancer_numeric",
    "any_neurological_numeric",
    "any_allergy_numeric",
    "bone_joint_problem_numeric",
    "lung_breathing_problems_numeric",
    "heart_circulation_problem_numeric",
    "digestive_problem_numeric",
    "diabetes_numeric",
    "polycystic_ovary_numeric",
    "skin_problems_numeric",
    "thyroid_disease_numeric"

  )
```

```{r broad categories somatic}
# Broad categories
dat_somatic_dx_broad_longer <- dat %>%
  select(
    
    all_of(exclude_broad_columns_from_pivoting),
    
    # Diagnosis
    anorexia_nervosa,
    bulimia_nervosa,
    binge_eating_disorder,
    purging_disorder,
    atypical_anorexia_nervosa
  ) %>%
  pivot_longer(
    -exclude_broad_columns_from_pivoting,
#    names_to = "Old_variable",
    values_to = "Disorder"
    ) %>%
  select(
    -name
  )

dat_somatic_dx_broad_longer %>%
  head()
```


```{r}
save.image(file = paste0(filepath_cleaned_data, "tables_040523.Rdata"))
```
