==> scripts/analyses/3.Demographics.Rmd <==
---
title: "Demographics"
author: "Christopher Huebel, Helena Davies, Dina Monssen"
date: "2022-08-24"
output:
word_document: default
html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Clear Global Environment
```{r clear global environment}
rm(list = ls(all.names = TRUE)) 
```

Read in file with path to ilovedata channel on Teams
Ensure that your credentials directory is correctly located
```{r Read in paths file}
source(file = "../credentials/paths.R")
```

Add the add_numeric function - used to convert character variables into numeric variables.
Add the sumscores function - used to generate sumscores
Add the package_check function - used to install and load dependencies
```{r Read in functions}
source(file = paste0(ilovedata_scripts, "functions/add_numeric.R"))
source(file = paste0(ilovedata_scripts, "functions/remove_duplicates.R"))
source(file = paste0(ilovedata_scripts, "functions/sumscores.R"))
source(file = paste0(ilovedata_scripts, "functions/package_check.R"))
source(file = paste0(ilovedata_scripts, "functions/imp_check.R"))
source(file = paste0(ilovedata_scripts, "functions/recode_check.R"))
```

Load packages

```{r Load packages}
library(tidyverse)
library(flextable)
library(officer)
library(gtsummary)
```

Set current date

```{r current date}
date <- Sys.Date()
```

Load data

```{r Load data}
load(file = paste0(filepath_cleaned_data, "tables_040523.Rdata"))
```

```{r}
names(dat_demographics)
```

# Demographics whole sample
```{r demographics whole sample}
summary_table_whole_sample <-
  dat_demographics %>%
  select(-mixed_eating_disorders_all) %>%
  gtsummary::tbl_summary(
        
        label = list(
          age ~ 
            "Age",
          age_missing_info ~ 
            "Age: missing info",
          sex ~
            "Female",
         gender ~
         "Gender",
        sexuality ~
          "Sexual orientation",
        dem.do_you_identify_as_transgender ~
          "Transgender",
        ethnicity ~
          "Ethnic origin",
        dem.highest_education_finegrained ~
          "Highest education level",
        dem.what_is_your_current_employment_status ~
          "Current employment status",
        relationship_status ~
          "Current marital/relationship Status",
        mhd.eating_disorder_received_treatment ~
            "Received treatment",
        dem.pack_year ~
          "Smoking: packs per year"
        ),
        
          type = list(
            age ~ 
              "continuous2",
            age_missing_info ~ 
              "categorical",
            sex ~
            "dichotomous",
            gender ~ 
            "categorical",
          sexuality ~
            "categorical",
          dem.do_you_identify_as_transgender ~
            "dichotomous",
          ethnicity ~
            "categorical",
          dem.highest_education_finegrained ~
            "categorical",
          dem.what_is_your_current_employment_status ~
            "categorical",
          relationship_status ~
            "categorical",          
          dem.pack_year ~ 
            "continuous2",
          mhd.eating_disorder_received_treatment ~
            "dichotomous"
          ),
           value = list(sex = "Female",
                   dem.do_you_identify_as_transgender = "Yes",
                   mhd.eating_disorder_received_treatment = 1),
        missing_text = "Not seen",
         digits = list(all_categorical() ~ c(0, 1),
                          all_dichotomous() ~ c(0, 1)),
             statistic = list(
                             
            all_categorical() ~ "{n} ({p}%)",
            
            all_continuous() ~ c("{mean} ({sd})", "{median} ({p25}, {p75})",  "{min} - {max}")))


summary_table_whole_sample
```

```{r summary_table table}
summary_table <-
  dat_demographics %>%
  gtsummary::tbl_summary(
        
        by =  mixed_eating_disorders_all,
        
        label = list(
           age ~ 
            "Age",
          sex~
            "Sex",
          gender ~
          "Gender",
        sexuality ~
          "Sexual orientation",
        dem.do_you_identify_as_transgender ~
          "Transgender",
        ethnicity ~
          "Ethnic origin",
        dem.highest_education_finegrained ~
          "Highest education level",
        dem.what_is_your_current_employment_status ~
          "Current employment status",
        relationship_status ~
          "Current marital/relationship status",
        mhd.eating_disorder_received_treatment ~
            "Received treatment",
        dem.pack_year ~
          "Smoking: pack per year"
        ),
        
          type = list(
            age ~ 
              "continuous2",
            sex ~
            "categorical",
            gender ~ 
            "categorical",
          sexuality ~
            "categorical",
          dem.do_you_identify_as_transgender ~
            "categorical",
          ethnicity ~
            "categorical",
          dem.highest_education_finegrained ~
            "categorical",
          dem.what_is_your_current_employment_status ~
            "categorical",
          relationship_status ~
            "categorical",          
          dem.pack_year ~ 
            "categorical",
          mhd.eating_disorder_received_treatment ~
            "categorical"
          ),
        value = list(sex = "Female",
                   dem.do_you_identify_as_transgender = "Yes",
                   mhd.eating_disorder_received_treatment = 1),
       missing_text = "Not seen",

          digits = list(all_categorical() ~ c(0, 1),
                          all_dichotomous() ~ c(0, 1)),
             statistic = list(
                             
            all_categorical() ~ "{n} ({p}%)",
            
            all_continuous() ~ c("{mean} ({sd})", "{median} ({p25}, {p75})",  "{min} - {max}")))

summary_table
```

```{r summary_table_long table}
summary_table_long <-
  dat_demographics_long %>%
  gtsummary::tbl_summary(
        
        by =  Disorder,
        
 label = list(
          age ~ 
          "Age",
          age_missing_info ~
            "Age: missing info",
          sex ~
            "Female",
          gender ~
          "Gender",
        sexuality ~
          "Sexual orientation",
        dem.do_you_identify_as_transgender ~
          "Transgender",
        ethnicity ~
          "Ethnic origin",
        dem.highest_education_finegrained ~
          "Highest education level",
        dem.what_is_your_current_employment_status ~
          "Current employment status",
        relationship_status ~
          "Current marital/relationship status",
        mhd.eating_disorder_received_treatment ~
            "Received treatment",
          dem.pack_year ~
          "Smoking: pack per year",
        dem.pack_year_missing_info ~
          "Smoking: missing info"
        ),
        
          type = list(
            age ~ 
             "continuous2",
            age_missing_info ~
              "categorical",
            sex ~
            "dichotomous",
            gender ~ 
            "categorical",
          sexuality ~
            "categorical",
          dem.do_you_identify_as_transgender ~
            "dichotomous",
          ethnicity ~
            "categorical",
          dem.highest_education_finegrained ~
            "categorical",
          dem.what_is_your_current_employment_status ~
            "categorical",
          relationship_status ~
            "categorical",          
          mhd.eating_disorder_received_treatment ~
            "dichotomous",
          dem.pack_year ~
            "continuous2",
          dem.pack_year_missing_info ~
            "categorical"
          ),
       value = list(sex = "Female",
                   dem.do_you_identify_as_transgender = "Yes",
                   mhd.eating_disorder_received_treatment = 1),
    missing_text = "Not seen",
 digits = list(all_categorical() ~ c(0, 1),
                          all_dichotomous() ~ c(0, 1)),
             statistic = list(
                             
            all_categorical() ~ "{n} ({p}%)",
            
            all_continuous() ~ c("{mean} ({sd})", "{median} ({p25}, {p75})",  "{min} - {max}")))

summary_table_long

```

Add bold labels to table
```{r summary_table with bold labels}
summary_table <- summary_table %>% 
  bold_labels() 
```

```{r summary_table_long with bold labels}
summary_table_long <- summary_table_long %>% 
  bold_labels() 
```

Save as a word document
```{r word properties}
sect_properties <- officer::prop_section(
  page_size = page_size(
    orient = "landscape",
    width = 8.3, height = 11.7
  ),
  type = "continuous",
  page_margins = page_mar()
)
```

```{r summary_table as word document}
summary_table %>%
  as_flex_table() %>%
  flextable::save_as_docx(
    path = "../../results/supplement/EDGI_demographics_040523.docx",
    pr_section = sect_properties
  )
```

```{r summary_table_long as word document}
summary_table_long %>%
  as_flex_table() %>%
  flextable::save_as_docx(
    path = "../../results/main_manuscript/EDGI_demographics_long_for_manuscript_040523.docx",
    pr_section = sect_properties
  )
```

```{r summary_table_long as word document}
summary_table_whole_sample %>%
  as_flex_table() %>%
  flextable::save_as_docx(
    path = "../../results/main_manuscript/EDGI_demographics_long_wholesample_for_manuscript_040523.docx",
    pr_section = sect_properties
  )
```
