==> scripts/analyses/4.Age.Rmd <==
---
title: "Demographics"
author: "Christopher Huebel, Helena Davies, Dina Monssen"
date: "2022-08-24"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Clear Global Environment

```{r clear global environment}
rm(list = ls(all.names = TRUE)) 
```

Read in file with path to data channel
```{r Read in file with path to data channel on teams}
source(file = "../credentials/paths.R")
```

Add the add_numeric function - used to convert character variables into numeric variables
Add the remove_duplicates function - used to deduplicate and remove NAs from IDs
Add the sumscores function - used to generate sumscores
Add the package_check function - used to install and load dependencies
```{r Read in functions}
source(file = paste0(ilovedata_scripts, "functions/add_numeric.R"))
source(file = paste0(ilovedata_scripts, "functions/remove_duplicates.R"))
source(file = paste0(ilovedata_scripts, "functions/sumscores.R"))
source(file = paste0(ilovedata_scripts, "functions/package_check.R"))
source(file = paste0(ilovedata_scripts, "functions/imp_check.R"))
source(file = paste0(ilovedata_scripts, "functions/recode_check.R"))
```

Load packages

```{r Load packages}
library(tidyverse)
library(officer)
library(flextable)
library(gtsummary)
```

Set current date

```{r current date}
date <- Sys.Date()
```

Load data

```{r Load data}
load(file = paste0(filepath_cleaned_data, "tables_040523.Rdata"))
```


```{r age_sex_distribution names}
names(age_sex_distribution)
```



```{r Check all NA participants}
age_sex_distribution %>%
  filter(
    is.na(sex)
  )
```

```{r Check sex variable}
age_sex_distribution %>%
  summarytools::freq(sex)
```

Check age variable
```{r}
age_sex_distribution %>%
  summarytools::freq(age)
```

```{r}
age_sex_distribution %>%
  filter(
    !is.na(sex)
  ) %>%
  filter(
    age != -666
  ) %>%
  summarytools::freq(
    age
  )
```

```{r Age distribution plot by sex}
png(file = "../../results/main_manuscript/EDGI_age_distribution_by_sex_for_manuscript_040523.png",
    width = 800,
    height = 500)

age_sex_distribution %>%
  ggplot(aes (
    x = age,
    fill = sex,
    alpha = 0.5)) + 
  geom_density() +
      labs(
      x = "Age [years]",
      y = "Percentage of total sample",
      fill = "Sex",
      title = element_blank(),
      color = "black"
    ) +
    theme(
      panel.grid.major.y = element_line(
        size = 0.5,
        linetype = "dashed",
        colour = "gray"
      ),
       axis.title.x = element_text(colour = "black", size = 20),
      axis.title.y = element_text(colour = "black", size = 20),
      axis.text.x = element_text(colour = "black", size = 15),
      axis.text.y = element_text(colour = "black", size = 15),
      axis.ticks.x = element_blank(),
      axis.ticks.y = element_blank(),
      panel.background = element_blank(),
      legend.key.size = unit(1, 'cm'), # change legend key size
        legend.key.height = unit(1, 'cm'), #change legend key height
        legend.key.width = unit(1, 'cm'), #change legend key width
        legend.title = element_text(size=14), #change legend title font size
        legend.text = element_text(size=14)) + # change legend text font size 
scale_alpha(guide = 'none') + scale_y_continuous(labels = scales::percent) +
  scale_fill_manual(values=c("#F68E3E", "#8b3582"))

dev.off()
```



```{r}
# Long format
exclude_columns_from_pivoting <-
  c(
    "age"
    )

age_sex_distribution_long <- dat %>%
  select(
    all_of(exclude_columns_from_pivoting),
         
        # Disorder
        anorexia_nervosa,
        bulimia_nervosa,
        binge_eating_disorder,
        purging_disorder,
    atypical_anorexia_nervosa
        ) %>%
  pivot_longer(
    -exclude_columns_from_pivoting,
#    names_to = "Old_variable",
    values_to = "Disorder"
    ) %>%
  select(
    -name
  )

age_sex_distribution_long
```





```{r summary_table_long table}
summary_table_long <-
  age_sex_distribution_long %>%
  gtsummary::tbl_summary(
        
        by =  Disorder,
        
        label = list(

        age ~
          "Age"
        ),
                
          type = list(
            age ~
          "categorical"),

          statistic = list(
                             
            all_categorical() ~ "{n} ({p}%)",
            
            all_continuous2() ~ c("{mean} ({sd})",
                                  "{median} ({p25}, {p75})",
                                  "{min}, {max}"
                                  )
                ), 
              
              missing = "always") 
```

Add bold labels to table
```{r summary_table_long with bold labels}
summary_table_long %>% 
  bold_labels() 
```

Save as a word document
```{r modify labls and run table}
summary_table_long %>% 
  modify_caption("**Table 1. Patient Characteristics**") %>%  
  modify_footnote(all_stat_cols() ~ "n(%) or Mean(SD)") %>% 
  bold_labels() %>% 
  add_n() %>% 
  modify_header(label ~ "**Variable**") 
```
