==> scripts/analyses/5.ED_diagnoses.Rmd <==
---
title: "Demographics"
author: "Christopher Huebel, Helena Davies, Dina Monssen"
date: "2022-08-24"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Clear Global Environment
```{r clear global environment}
rm(list = ls(all.names = TRUE)) 
```

Read in file with path to data channel
```{r Read in file with path to data channel on teams}
source(file = "../credentials/paths.R")
```

Add the add_numeric function - used to convert character variables into numeric variables
Add the remove_duplicates function - used to deduplicate and remove NAs from IDs
Add the sumscores function - used to generate sumscores
Add the package_check function - used to install and load dependencies
Add the imp_check function - used to check variables for implausible values
```{r Read in functions}
source(file = paste0(ilovedata_scripts, "functions/add_numeric.R"))
source(file = paste0(ilovedata_scripts, "functions/remove_duplicates.R"))
source(file = paste0(ilovedata_scripts, "functions/sumscores.R"))
source(file = paste0(ilovedata_scripts, "functions/package_check.R"))
source(file = paste0(ilovedata_scripts, "functions/imp_check.R"))
source(file = paste0(ilovedata_scripts, "functions/recode_check.R"))
```

Load packages
```{r Load packages}
library(flextable)
library(officer)
library(gtsummary)
library(UpSetR)
library(tidyverse)
```

Set current date
```{r current date}
date <- Sys.Date()
```

Load data
```{r Load data}
load(file = paste0(filepath_cleaned_data, "tables_040523.Rdata"))
```

```{r}
names(selfreport_algo_sex)
```

# Table of frequency of eating disorder
```{r table freq EDs}
# Create OSFED variable
dat <- dat %>%
  mutate(OSFED_EDNOS_numeric =
           case_when(mhd.feeding_eating_disorder_numeric == 1 ~ 1,
                     mhd.feeding_eating_disorder_numeric == 0 ~ 0,
                     TRUE ~ NA_real_))

dat_eds_binary_numeric <- dat %>%
  select(
    "Anorexia nervosa" = anorexia_nervosa_subtype_numeric,
    "Bulimia nervosa" = bulimia_nervosa_numeric,
    "Binge-eating disorder" = binge_eating_disorder_numeric,
    "Purging disorder" = purging_disorder_numeric,
    "Atypical anorexia nervosa" = atypical_anorexia_nervosa_numeric,
    "Atypical binge-eating disorder" = atypical_binge_eating_disorder_numeric,
    "Atypical bulimia nervosa" = atypical_bulimia_nervosa_numeric,
    "Night-eating syndrome" = night_eating_syndrome_numeric,
    "Pica" = pica_numeric,
    "Avoidant/restrictive food intake disorder" = avoidant_restrictive_food_intake_disorder_numeric,
    "Rumination disorder" = rumination_disorder_numeric,
    "OSFED/EDNOS" = OSFED_EDNOS_numeric
  ) %>%
  gtsummary::tbl_summary(missing = "ifany")

dat_eds_binary_numeric
```

# Table of frequency of eating disorder by sex
```{r table freq EDs}
dat_eds_binary_sex_numeric <- dat %>%
  select(
    "Anorexia nervosa" = anorexia_nervosa_subtype_numeric,
    "Bulimia nervosa" = bulimia_nervosa_numeric,
    "Binge-eating disorder" = binge_eating_disorder_numeric,
    "Purging disorder" = purging_disorder_numeric,
    "Atypical anorexia nervosa" = atypical_anorexia_nervosa_numeric,
    "Atypical binge-eating disorder" = atypical_binge_eating_disorder_numeric,
    "Atypical bulimia nervosa" = atypical_bulimia_nervosa_numeric,
    "Night-eating syndrome" = night_eating_syndrome_numeric,
    "Pica" = pica_numeric,
    "Avoidant/restrictive food intake disorder" = avoidant_restrictive_food_intake_disorder_numeric,
    "Rumination disorder" = rumination_disorder_numeric,
    "OSFED/EDNOS" = OSFED_EDNOS_numeric,
    sex
  ) %>%
  gtsummary::tbl_summary(missing = "no",
                         by = sex,
                         digits = ~ c(0, 1))


dat_eds_binary_sex_numeric
```

```{r upset plot, fig.height=6, fig.width=7}
# Create numeric binary variables needed
dat <- dat %>% 
  mutate(ed.DSM5_AN_binge_purge_binary_numeric = 
    case_when(ed.DSM5_AN_binge_purge_binary == "DSM-5 AN binge eating/purging" ~ 1,
         TRUE ~ 0
    )
  )

dat <- dat %>% 
  mutate(ed.DSM5_AN_restricting_binary_numeric = 
    case_when(ed.DSM5_AN_restricting_binary == "DSM-5 AN restricting" ~ 1,
        TRUE ~ 0
    )
  )


dat <- dat %>% 
  mutate(anorexia_nervosa_subtype_unknown_numeric = 
    case_when(anorexia_nervosa_subtype_numeric == "Unknown subtype" ~ 1,
         
              TRUE~ 0
    )
  )


dat_upset <- dat %>% 
   select(
     "ANR" = ed.DSM5_AN_restricting_binary_numeric,
     "ANBP" = ed.DSM5_AN_binge_purge_binary_numeric,
     "AN" = anorexia_nervosa_subtype_unknown_numeric,
    "BN" = bulimia_nervosa_numeric,
    "BED" = binge_eating_disorder_numeric,
    "PUR" = purging_disorder_numeric,
    "AAN" = atypical_anorexia_nervosa_numeric,
    "ABED" = atypical_binge_eating_disorder_numeric,
    "ABN" = atypical_bulimia_nervosa_numeric,
    "NES" = night_eating_syndrome_numeric,
    "Pica" = pica_numeric,
    "ARFID" = avoidant_restrictive_food_intake_disorder_numeric,
    "RUM" = rumination_disorder_numeric,
    "OSFED" = OSFED_EDNOS_numeric,
   ) %>%
  as.data.frame()

png(file = "../../results/main_manuscript/EDGI_upsetR_plot_for_manuscript_040523.png",
    width = 1000,
    height = 900)

# text.scale can either take a universal scale in the form of an integer, or a vector of specific scales in the format: c(intersection size title, intersection size tick labels, set size title, set size tick labels, set names, numbers above bars).

UpSetR::upset(dat_upset,
              sets = c("AN",
                      "ANR",
                      "ANBP",
    "BN",
    "BED",
    "PUR",
    "AAN" 
   # "Atypical binge-eating disorder",
   # "Atypical bulimia nervosa",
   #  "Night-eating syndrome",
   # "Pica",
   # "Avoidant/restrictive food intake disorder",
   # "Rumination disorder"
   ),
        nintersects = 27, # Limit to intersections with 20 or more participants
    order.by = "freq",
   text.scale = c(2,2,2,2,2,2)
   
    ) 


dev.off()

png(file = "../../results/supplement/EDGI_upsetR_plot_for_supplement_040523.png",
    width = 1200,
    height = 900)

UpSetR::upset(dat_upset,
              sets = c("AN",
                      "ANR",
                      "ANBP",
    "BN",
    "BED",
    "PUR",
    "AAN" ,
   # "Atypical binge-eating disorder",
   # "Atypical bulimia nervosa",
    "NES",
   # "Pica",
    "ARFID",
   "OSFED" 
   # "Rumination disorder"
   ),
        nintersects = 71, # Limit to intersections with 5 or more participants
    order.by = "freq",
    text.scale = c(2,2,2,2,2,1.65)
   
    ) 


dev.off()
```

```{r modify labls and run table}
dat_eds_binary_sex_numeric_mod <- dat_eds_binary_sex_numeric %>% 
  bold_labels() 
```

Save as a word document
```{r word properties}
sect_properties <- officer::prop_section(
  page_size = page_size(
    orient = "landscape",
    width = 8.3, height = 11.7
  ),
  type = "continuous",
  page_margins = page_mar()
)
```

```{r summary_table_long as word document}
dat_eds_binary_sex_numeric_mod %>%
  as_flex_table() %>%
  flextable::save_as_docx(
    path = "../../results/main_manuscript/EDGI_diagnosis_by_sex_for_manuscript_040523.docx",
    pr_section = sect_properties
  )
```
