==> scripts/analyses/6.Somatic_dx.Rmd <==
---
title: "Demographics"
author: "Christopher Huebel, Helena Davies, Dina Monssen"
date: "2022-08-24"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Clear Global Environment

```{r clear global environment}
rm(list = ls(all.names = TRUE)) 
```

Read in file with path to data channel
```{r Read in file with path to data channel on teams}
source(file = "../credentials/paths.R")
```

Add the add_numeric function - used to convert character variables into numeric variables
Add the remove_duplicates function - used to deduplicate and remove NAs from IDs
Add the sumscores function - used to generate sumscores
Add the package_check function - used to install and load dependencies
Add the imp_check function - used to check variables for implausible values
```{r Read in functions}
source(file = paste0(ilovedata_scripts, "functions/add_numeric.R"))
source(file = paste0(ilovedata_scripts, "functions/remove_duplicates.R"))
source(file = paste0(ilovedata_scripts, "functions/sumscores.R"))
source(file = paste0(ilovedata_scripts, "functions/package_check.R"))
source(file = paste0(ilovedata_scripts, "functions/imp_check.R"))
source(file = paste0(ilovedata_scripts, "functions/recode_check.R"))
```

Load packages

```{r Load packages}
library(tidyverse)
library(flextable)
library(officer)
library(gtsummary)
```

Set current date

```{r current date}
date <- Sys.Date()
```

Load data

```{r Load data}
load(file = paste0(filepath_cleaned_data, "tables_040523.Rdata"))
```


```{r}
names(dat_somatic_dx)
```


```{r summary_table table}
summary_table <-
  dat_somatic_dx %>%
  gtsummary::tbl_summary(
        
        by =  mixed_eating_disorders_all,
        
       label = list(
        dem.epilepsy_or_convulsions ~
          "Epilepsy or Convulsions",
        dem.migraines ~
          "Migraines",
        dem.multiple_sclerosis ~
          "Multiple Sclerosis",
        dem.parkinsons_disease ~
          "Parkinson's Disease",
        dem.severe_memory_loss ~
          "Severe Memory Loss",
        dem.hay_fever ~
          "Hay Fever",
        dem.drug_allergy ~                                                 
          "Drug Allergy",                                               
        dem.food_allergy ~
          "Food Allergy",
        dem.other_allergy ~
          "Other Allergies",
        dem.osteoporosis ~
          "Osteoporosis",
        dem.osteoarthritis ~
          "Osteoarthriris",
        dem.rheumatoid_arthritis ~
          "Rheumatoid Arthritis",
        dem.other_arthritis ~
          "Any Other Arthritis",
        dem.asthma ~
          "Asthma",
        dem.emphysema_or_chronic_bronchitis ~
          "Emphysema or Chronic Bronchitis",
        dem.heart_attack_or_angina ~
          "Heart Attack or Angina",
        dem.high_blood_cholesterol ~
          "High Cholesterol",
        dem.high_blood_pressure ~   
          "High Blood Pressure",
        dem.atrial_fibrillation ~
          "Atrial Fibrillation",
        dem.stroke ~
          "Stroke",
        dem.crohns_disease ~
          "Crohn's Disease",
        dem.ulcerative_colitis ~
          "Ulcerative Colitis",
        dem.coeliac_disease ~
          "Coeliac Disease",
        dem.diabetes_type_1_early_onset ~ 
          "Diabetes type 1 (early onset)",
        dem.diabetes_type_1_late_onset ~ 
          "Diabetes type 1 (late onset)",
        dem.diabetes_type_2_late_onset  ~
          "Diabetes type 2 (late onset)",
        dem.pain_due_to_diabetes ~
          "Diabetic-related pain",
         dem.pain_due_to_virus ~
            "Virus-related pain (post herpetic neuralgia)",
        dem.breast_cancer ~
          "Breast Cancer",
        dem.lung_cancer ~
          "Lung Cancer",
        dem.stomach_cancer ~
          "Stomach Cancer",
        dem.colon_cancer ~
          "Colon Cancer",
        dem.uterus_cancer ~
          "Uterus Cancer",
        dem.prostate_cancer ~
          "Prostate Cancer",
        dem.psoriasis ~
          "Psoriasis",
        dem.vitiligo ~
          "Vitiligo",
        dem.eczema ~
          "Eczema",
        dem.thyroid_disease ~
          "Thyroid Disease"
        ),
                
          type = list(
            dem.epilepsy_or_convulsions ~
          "dichotomous",
        dem.migraines ~
          "dichotomous",
        dem.multiple_sclerosis ~
          "dichotomous",
        dem.parkinsons_disease ~
          "dichotomous",
        dem.severe_memory_loss ~
          "dichotomous",
        dem.hay_fever ~
          "dichotomous",
        dem.drug_allergy ~                                                 
          "dichotomous",                                               
        dem.food_allergy ~
          "dichotomous",
        dem.other_allergy ~
          "dichotomous",
        dem.osteoporosis ~
          "dichotomous",
        dem.osteoarthritis ~
          "dichotomous",
        dem.rheumatoid_arthritis ~
          "dichotomous",
        dem.other_arthritis ~
          "dichotomous",
        dem.asthma ~
          "dichotomous",
        dem.emphysema_or_chronic_bronchitis ~
          "dichotomous",
        dem.heart_attack_or_angina ~
          "dichotomous",
        dem.high_blood_cholesterol ~
          "dichotomous",
        dem.high_blood_pressure ~   
          "dichotomous",
        dem.atrial_fibrillation ~
          "dichotomous",
        dem.stroke ~
          "dichotomous",
        dem.crohns_disease ~
          "dichotomous",
        dem.ulcerative_colitis ~
          "dichotomous",
        dem.coeliac_disease ~
          "dichotomous",
       dem.diabetes_type_1_early_onset ~ 
          "dichotomous",
        dem.diabetes_type_1_late_onset ~ 
          "dichotomous",
        dem.diabetes_type_2_late_onset  ~
          "dichotomous",
        dem.pain_due_to_diabetes ~
          "dichotomous",
       dem.pain_due_to_virus ~
            "dichotomous",
        dem.breast_cancer ~
          "dichotomous",
        dem.lung_cancer ~
          "dichotomous",
        dem.stomach_cancer ~
          "dichotomous",
        dem.colon_cancer ~
          "dichotomous",
        dem.uterus_cancer ~
          "dichotomous",
        dem.prostate_cancer ~
          "dichotomous",
        dem.psoriasis ~
          "dichotomous",
        dem.vitiligo ~
          "dichotomous",
        dem.eczema ~
          "dichotomous",
        dem.thyroid_disease ~
          "dichotomous"
        ),
           value =
       list(
       dem.epilepsy_or_convulsions = "Epilepsy or convulsions",
       dem.breast_cancer = "Breast cancer",
       dem.lung_cancer = "Lung cancer",
       dem.stomach_cancer = "Stomach cancer",
       dem.colon_cancer = "Colon cancer",
       dem.uterus_cancer =  "Uterus cancer",
       dem.prostate_cancer =  "Prostate cancer",
       dem.epilepsy_or_convulsions = "Epilepsy or convulsions",
       dem.migraines = "Migraines",
        dem.multiple_sclerosis = "Multiple sclerosis",
        dem.parkinsons_disease = "Parkinson's disease",
        dem.brain_tumour = "Brain tumour",
        dem.severe_memory_loss = "Severe memory loss",
        dem.hay_fever = "Hay fever",
        dem.drug_allergy = "Drug allergy",
        dem.food_allergy = "Food allergy",
        dem.other_allergy = "Other allergy",
        dem.ankylosing_spondylitis = "Ankylosing spondylitis",
        dem.osteoporosis = "Osteoporosis",
        dem.osteoarthritis = "Osteoarthritis",
        dem.hypermobility = "Hyper-mobility",
        dem.rheumatoid_arthritis = "Rheumatoid arthritis",
        dem.other_arthritis = "Other arthritis",
        dem.emphysema_or_chronic_bronchitis = "Emphysema or chronic bronchitis",
        dem.heart_attack_or_angina ="Heart attack or angina",
        dem.high_blood_cholesterol = "High blood cholesterol",
        dem.high_blood_pressure = "High blood pressure",
        dem.atrial_fibrillation = "Atrial fibrillation",
        dem.stroke = "Stroke",
        dem.pots = "Paroxysmal tachycardia syndrome (PoTS)",
        dem.crohns_disease = "Crohn's disease",
        dem.ulcerative_colitis = "Ulcerative colitis",
        dem.coeliac_disease = "Coeliac disease",
        dem.diabetes_type_1_early_onset = "Diabetes type 1 (early onset)",
        dem.diabetes_type_1_late_onset =  "Diabetes type 1 (late onset)",
        dem.diabetes_type_2_late_onset  = "Diabetes type 2 (late onset)",
        dem.pain_due_to_diabetes = "Pain due to diabetes (diabetic neuropathy)",
        dem.pcos = "Polycystic ovary syndrome (PCOS)",
        dem.psoriasis = "Psoriasis",
        dem.eczema = "Eczema",
        dem.vitiligo = "Vitiligo",
        dem.pain_due_to_virus = "Pain due to virus",
        dem.thyroid_disease = "Thyroid disease",
       dem.asthma = "Asthma"
  ),
          
      digits = list(all_categorical() ~ c(0, 1),
                          all_dichotomous() ~ c(0, 1)),
             statistic = list(
                             
            all_categorical() ~ "{n} ({p}%)"))


```


```{r summary_table_long table}
summary_table_long <-
  dat_somatic_dx_longer %>%
  gtsummary::tbl_summary(
        
        by =  Disorder,
        
       label = list(
        dem.epilepsy_or_convulsions ~
          "Epilepsy or Convulsions",
        dem.migraines ~
          "Migraines",
        dem.multiple_sclerosis ~
          "Multiple Sclerosis",
        dem.parkinsons_disease ~
          "Parkinson's Disease",
        dem.severe_memory_loss ~
          "Severe Memory Loss",
        dem.hay_fever ~
          "Hay Fever",
        dem.drug_allergy ~                                                 
          "Drug Allergy",                                               
        dem.food_allergy ~
          "Food Allergy",
        dem.other_allergy ~
          "Other Allergies",
        dem.osteoporosis ~
          "Osteoporosis",
        dem.osteoarthritis ~
          "Osteoarthriris",
        dem.rheumatoid_arthritis ~
          "Rheumatoid Arthritis",
        dem.other_arthritis ~
          "Any Other Arthritis",
        dem.asthma ~
          "Asthma",
        dem.emphysema_or_chronic_bronchitis ~
          "Emphysema or Chronic Bronchitis",
        dem.heart_attack_or_angina ~
          "Heart Attack or Angina",
        dem.high_blood_cholesterol ~
          "High Cholesterol",
        dem.high_blood_pressure ~   
          "High Blood Pressure",
        dem.atrial_fibrillation ~
          "Atrial Fibrillation",
        dem.stroke ~
          "Stroke",
        dem.crohns_disease ~
          "Crohn's Disease",
        dem.ulcerative_colitis ~
          "Ulcerative Colitis",
        dem.coeliac_disease ~
          "Coeliac Disease",
         dem.diabetes_type_1_early_onset ~
          "Diabetes type 1 (early onset)",
        dem.diabetes_type_1_late_onset ~
          "Diabetes type 1 (late onset)",
        dem.diabetes_type_2_late_onset ~
          "Diabetes type 2 (late onset)",
       dem.pain_due_to_diabetes ~
          "Diabetic-related Pain",
        dem.pain_due_to_virus ~
            "Virus-related pain (Post herpetic neuralgia)",
        dem.breast_cancer ~
          "Breast Cancer",
        dem.lung_cancer ~
          "Lung Cancer",
        dem.stomach_cancer ~
          "Stomach Cancer",
        dem.colon_cancer ~
          "Colon Cancer",
        dem.uterus_cancer ~
          "Uterus Cancer",
        dem.prostate_cancer ~
          "Prostate Cancer",
        dem.psoriasis ~
          "Psoriasis",
        dem.vitiligo ~
          "Vitiligo",
        dem.eczema ~
          "Eczema",
        dem.thyroid_disease ~
          "Thyroid Disease"
        ),
                
          type = list(
            dem.epilepsy_or_convulsions ~
          "dichotomous",
        dem.migraines ~
          "dichotomous",
        dem.multiple_sclerosis ~
          "dichotomous",
        dem.parkinsons_disease ~
          "dichotomous",
        dem.severe_memory_loss ~
          "dichotomous",
        dem.hay_fever ~
          "dichotomous",
        dem.drug_allergy ~                                                 
          "dichotomous",                                               
        dem.food_allergy ~
          "dichotomous",
        dem.other_allergy ~
          "dichotomous",
        dem.osteoporosis ~
          "dichotomous",
        dem.osteoarthritis ~
          "dichotomous",
        dem.rheumatoid_arthritis ~
          "dichotomous",
        dem.other_arthritis ~
          "dichotomous",
        dem.asthma ~
          "dichotomous",
        dem.emphysema_or_chronic_bronchitis ~
          "dichotomous",
        dem.heart_attack_or_angina ~
          "dichotomous",
        dem.high_blood_cholesterol ~
          "dichotomous",
        dem.high_blood_pressure ~   
          "dichotomous",
        dem.atrial_fibrillation ~
          "dichotomous",
        dem.stroke ~
          "dichotomous",
        dem.crohns_disease ~
          "dichotomous",
        dem.ulcerative_colitis ~
          "dichotomous",
        dem.coeliac_disease ~
          "dichotomous",
        dem.diabetes_type_1_early_onset ~
          "dichotomous",
        dem.diabetes_type_1_late_onset ~
          "dichotomous",
        dem.diabetes_type_2_late_onset ~
          "dichotomous",
       dem.pain_due_to_diabetes ~
          "dichotomous",
       dem.pain_due_to_virus ~
            "dichotomous",
        dem.breast_cancer ~
          "dichotomous",
        dem.lung_cancer ~
          "dichotomous",
        dem.stomach_cancer ~
          "dichotomous",
        dem.colon_cancer ~
          "dichotomous",
        dem.uterus_cancer ~
          "dichotomous",
        dem.prostate_cancer ~
          "dichotomous",
        dem.psoriasis ~
          "dichotomous",
        dem.vitiligo ~
          "dichotomous",
        dem.eczema ~
          "dichotomous",
        dem.thyroid_disease ~
          "dichotomous"
        ),
       
       value =
       list(
       dem.epilepsy_or_convulsions = "Epilepsy or convulsions",
       dem.lung_cancer = "Lung cancer",
       dem.breast_cancer = "Breast cancer",
       dem.stomach_cancer = "Stomach cancer",
       dem.colon_cancer = "Colon cancer",
       dem.uterus_cancer =  "Uterus cancer",
       dem.prostate_cancer =  "Prostate cancer",
       dem.epilepsy_or_convulsions = "Epilepsy or convulsions",
       dem.migraines = "Migraines",
        dem.multiple_sclerosis = "Multiple sclerosis",
        dem.parkinsons_disease = "Parkinson's disease",
        dem.brain_tumour = "Brain tumour",
        dem.severe_memory_loss = "Severe memory loss",
        dem.hay_fever = "Hay fever",
        dem.drug_allergy = "Drug allergy",
        dem.food_allergy = "Food allergy",
        dem.other_allergy = "Other allergy",
        dem.ankylosing_spondylitis = "Ankylosing spondylitis",
        dem.osteoporosis = "Osteoporosis",
        dem.osteoarthritis = "Osteoarthritis",
        dem.hypermobility = "Hyper-mobility",
        dem.rheumatoid_arthritis = "Rheumatoid arthritis",
        dem.other_arthritis = "Other arthritis",
        dem.emphysema_or_chronic_bronchitis = "Emphysema or chronic bronchitis",
        dem.heart_attack_or_angina ="Heart attack or angina",
        dem.high_blood_cholesterol = "High blood cholesterol",
        dem.high_blood_pressure = "High blood pressure",
        dem.atrial_fibrillation = "Atrial fibrillation",
        dem.stroke = "Stroke",
        dem.pots = "Paroxysmal tachycardia syndrome (PoTS)",
        dem.crohns_disease = "Crohn's disease",
        dem.ulcerative_colitis = "Ulcerative colitis",
        dem.coeliac_disease = "Coeliac disease",
        dem.diabetes_type_1_early_onset = "Diabetes type 1 (early onset)",
        dem.diabetes_type_1_late_onset =  "Diabetes type 1 (late onset)",
        dem.diabetes_type_2_late_onset  = "Diabetes type 2 (late onset)",
        dem.pain_due_to_diabetes = "Pain due to diabetes (diabetic neuropathy)",
        dem.pcos = "Polycystic ovary syndrome (PCOS)",
        dem.psoriasis = "Psoriasis",
        dem.eczema = "Eczema",
        dem.vitiligo = "Vitiligo",
        dem.pain_due_to_virus = "Pain due to virus",
        dem.thyroid_disease = "Thyroid disease",
        dem.asthma = "Asthma"),
       
      digits = list(all_dichotomous() ~ c(0, 1)),
             statistic = list(
                             
            all_categorical() ~ "{n} ({p}%)"))

summary_table_long
```

```{r summary_table_long broad table}
summary_table_broad_long <-
  dat_somatic_dx_broad_longer %>%
  gtsummary::tbl_summary(
        
        by =  Disorder,
        
       label = list(
        any_cancer_numeric ~ "Cancer",
    any_neurological_numeric ~ "Nervous system problem/s",
    any_allergy_numeric ~ "Allergy",
    bone_joint_problem_numeric ~ "Bone and/or joint problem/s",
    lung_breathing_problems_numeric ~ "Lung and/or breathing problem/s",
    heart_circulation_problem_numeric ~ "Heart and/or circulation problem/s",
    digestive_problem_numeric ~ "Digestive problem/s",
    diabetes_numeric ~ "Diabetes",
    polycystic_ovary_numeric ~ "Polycystic ovary syndrome",
    skin_problems_numeric ~ "Skin problem/s",
    thyroid_disease_numeric ~ "Thyroid disease"
        ),
                
          type = list(
    any_cancer_numeric   ~ "dichotomous",
    any_neurological_numeric ~ "dichotomous",
    any_allergy_numeric ~ "dichotomous",
    bone_joint_problem_numeric ~ "dichotomous",
    lung_breathing_problems_numeric ~ "dichotomous",
    heart_circulation_problem_numeric ~ "dichotomous",
    digestive_problem_numeric ~ "dichotomous",
    diabetes_numeric ~ "dichotomous",
    polycystic_ovary_numeric ~ "dichotomous",
    skin_problems_numeric ~ "dichotomous",
    thyroid_disease_numeric ~ "dichotomous"
        ),
       
       value =
       list(
       any_cancer_numeric ~ 1,
    any_neurological_numeric ~ 1,
    any_allergy_numeric ~ 1,
    bone_joint_problem_numeric ~ 1,
    lung_breathing_problems_numeric ~ 1,
    heart_circulation_problem_numeric ~ 1,
    digestive_problem_numeric ~ 1,
    diabetes_numeric ~ 1,
    polycystic_ovary_numeric ~ 1,
    skin_problems_numeric ~ 1,
    thyroid_disease_numeric ~ 1),
       
      digits = list(all_dichotomous() ~ c(0, 1)))


summary_table_broad_long
```

Add bold labels to table
```{r summary_table with bold labels}
summary_table %>% 
  bold_labels() 
```

```{r summary_table_long with bold labels}
summary_table_long %>% 
  bold_labels() 
```

Save as a word document
```{r word properties}
sect_properties <- officer::prop_section(
  page_size = 
    page_size(
      orient = "landscape",
      width = 8.3,
      height = 11.7
      ),
  type = "continuous",
  page_margins = page_mar()
)
```

```{r summary_table as word document}
summary_table %>%
  as_flex_table() %>%
  flextable::save_as_docx(
    path = "../../results/supplement/EDGI_somatic_dx_040523.docx",
    pr_section = sect_properties
  )
```

```{r}
summary_table_broad_long %>%
  as_flex_table() %>%
  flextable::save_as_docx(
    path = "../../results/main_manuscript/EDGI_somatic_dx_long_for_manuscript_040523.docx",
    pr_section = sect_properties
  )
```

```{r summary_table_long as word document}
summary_table_long %>%
  as_flex_table() %>%
  flextable::save_as_docx(
    path = "../../results/supplement/EDGI_somatic_dx_long_for_supplement_040523.docx",
    pr_section = sect_properties
  )
```
